+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - added `particleInjected` event
-  - fired when a Particle Accelerator injects a particle into the world
+  - fired when a particle is injected into the Particle Accelerator
   - can be canceled to prevent the particle from being injected
   - holds utility methods for disabling special interactions for spawning nether and end portals
 - added `particleCollided` event
   - fired when a particle collides with another particle
   - can be canceled to consume the particles without an output
+- added `particleExited` event
+  - fired when a particle exits the Particle Accelerator
 - renamed `position` to `pos` and `blockEntity` to `soulCollector` in `soulCollection` event
 
 ## [0.2.0] - 2026-02-06